contractor and legalentity complete

+++ online-courses/internal/service/listener_service.go
@@ -35,7 +35,7 @@ func NewListenerService(db database.DB,
 	}
 }
 
-func (l *listenerService) CreateFullListener(ctx context.Context, models *dto.CreateListenerDTO) error {
+func (l *listenerService) CreateFullListener(ctx context.Context, models *dto.CreateListenerDTO, idLegalEntity, idContractor *uuid.UUID) error {
 
 	tx, err := l.db.BeginTx(ctx)
 	if err != nil {
@@ -47,8 +47,9 @@ func (l *listenerService) CreateFullListener(ctx context.Context, models *dto.Cr
 	var passportID *uuid.UUID = nil
 
 	if models.Passport != (dto.PassportDTO{}) {
-		passportID := uuid.New()
-		passport, err := mapping.MapPassportToEntity(models.Passport, passportID)
+		newID := uuid.New()
+		passportID = &newID
+		passport, err := mapping.MapPassportToEntity(models.Passport, *passportID)
 		if err != nil {
 			return err
 		}
@@ -103,6 +104,8 @@ func (l *listenerService) CreateFullListener(ctx context.Context, models *dto.Cr
 		ID_RegAddress:        registrationID,
 		ID_EducationListener: educationID,
 		ID_PlaceWork:         placeWorkID,
+		ID_LegalEntity:       idLegalEntity,
+		ID_Contractor:        idContractor,
 	}
 
 	if err = l.listenerRepo.CreateInTx(ctx, tx, listener, optionalID); err != nil {
@@ -171,7 +174,40 @@ func (l *listenerService) ReadFullListener(ctx context.Context, id uuid.UUID) (*
 		result.Passport = *passport
 	}
 
-	return mapping.MapListenerEntityToDTO(result), nil
+	contractor := dto.ContractorCreateDTO{
+		Contractor: dto.ContractorDTO{
+			ID_Contractor: result.Contractor.ID_Contractor,
+			FirstName:     result.Contractor.FirstName,
+			SecondName:    result.Contractor.SecondName,
+			MiddleName:    result.Contractor.MiddleName,
+			Contact_phone: result.Contractor.Contact_phone,
+			Email:         result.Contractor.Email,
+		},
+		Passport: dto.PassportDTO{
+			PlaceBirth:    result.Contractor.Passport.PlaceBirth,
+			Citizenship:   result.Contractor.Passport.Citizenship,
+			Gender:        result.Contractor.Passport.Gender,
+			Seria:         result.Contractor.Passport.Seria,
+			Number:        result.Contractor.Passport.Number,
+			PassportGiven: result.Contractor.Passport.PassportGiven,
+			DateGiven:     result.Contractor.Passport.DateGiven.String(),
+			Code:          result.Contractor.Passport.Code,
+		},
+		RegAddress: dto.RegistrationAddressDTO{
+			MailIndex: result.Contractor.RegistrationAddress.MailIndex,
+			Region:    result.Contractor.RegistrationAddress.Region,
+			City:      result.Contractor.RegistrationAddress.City,
+			Street:    result.Contractor.RegistrationAddress.Street,
+			House:     result.Contractor.RegistrationAddress.House,
+			Building:  result.Contractor.RegistrationAddress.Building,
+			Apartment: result.Contractor.RegistrationAddress.Apartment,
+		},
+	}
+
+	listener := mapping.MapListenerEntityToDTO(result)
+	listener.Contractor = &contractor
+
+	return listener, nil
 }
 
 func (l *listenerService) UpdateListener(ctx context.Context, models *dto.CreateListenerDTO, id uuid.UUID) error {
@@ -254,8 +290,10 @@ func (l *listenerService) DeleteListener(ctx context.Context, id uuid.UUID) erro
 		return err
 	}
 
-	if err = l.passportRepo.DeleteInTx(ctx, tx, *ids.ID_Passport); err != nil {
-		return err
+	if ids.ID_Passport != nil {
+		if err = l.passportRepo.DeleteInTx(ctx, tx, *ids.ID_Passport); err != nil {
+			return err
+		}
 	}
 
 	if err = l.registrationRepo.DeleteInTx(ctx, tx, ids.ID_RegAddress); err != nil {
